✨ refactor: Remove unnecessary page title elements
This commit removes the unnecessary page title `<h1>` elements from the HTML files. The page title information is already provided in the breadcrumb navigation, so the redundant `<h1>` tags are being removed to simplify the markup and improve the overall structure of the pages.

The changes were made to the following files:

- `.github/assets/custom_template.html`
- `docs/index.html`
- `docs/src-local/dye-injection.html`
- `docs/testCases/LidDrivenCavity-Newtonian-dyeInjection.html`
- `docs/postProcess/getData-LidDriven.html`

This change helps to reduce visual clutter and makes the pages more concise and easier to read.