🔖 docs(title): Update page titles to reflect project name
This commit updates the page titles in the documentation to use a more descriptive and project-specific title "Lid driven cavity using dye injection" instead of the generic "CoMPhy-Lab Documentation".

The changes include updating the titles in the following files:
- `docs/src-local/dye-injection.h.html`
- `docs/reset_install_requirements.sh.html`
- `docs/postProcess/getData-LidDriven.c.html`
- `docs/postProcess/2-LidDrivenCavity-Newtonian-dyeInjection.py.html`
- `docs/index.html`
- `docs/testCases/LidDrivenCavity-Newtonian-dyeInjection.c.html`

This change helps to better reflect the focus of the project and provide more context for users navigating the documentation.